kselftests: split into 4.12 and -next
The current recipe for kselftests-4.12 is based on 4.12 plus
patches from Shuah Khan's next branch, plus a few other
patches that are relevant that have not made it into the
tree.

This split produces two recipes:
* kselftests 4.12
* kselftests next

The former is kselftests as they are found in mainline kernel,
specifically version 4.12, and reproduces what was already in
kselftests recipe. The latter tracks Shuah's next branch,
currently at 4.13-rc1, and is updated more often.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>