kselftests: split recipe into generic form
The current recipe for kselftests-4.12 is based on 4.12 plus
patches from Shuah Khan's next branch, plus a few other
patches that are relevant that have not made it into the
tree.

This split separates between the relevant bits of 4.12 and
the core that can be reused with other future recipes.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>